requirements.txt
pytelegrambotapi>=4.0.0
gspread>=6.2.1
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
requests>=2.28.0 